chore(interface): functions/srcにinterfaceを移動
