+++ functions/src/interfaces/article-with-author.ts
@@ -1,0 +1,6 @@
+import { Article } from './article';
+import { UserData } from './user';
+
+export interface ArticleWithAuthor extends Article {
+  author: UserData;
+}
