+++ src/app/shared/article/article.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit, Input } from '@angular/core';
-import { ArticleWithAuthor } from 'src/app/interfaces/article';
+import { ArticleWithAuthor } from 'functions/src/interfaces/article-with-author';
 
 @Component({
   selector: 'app-article',
