+++ tsconfig.json
@@ -11,10 +11,10 @@
     "moduleResolution": "node",
     "importHelpers": true,
     "target": "es2015",
-    "lib": [
-      "es2018",
-      "dom"
-    ]
+    "lib": ["es2018", "dom"],
+    "paths": {
+      "@interfaces/*": ["functions/src/interfaces/*"]
+    }
   },
   "angularCompilerOptions": {
     "fullTemplateTypeCheck": true,
